Float to Double

+++ src/main/java/com/ybigtaconference/gridtrading/service/Params.java
@@ -10,7 +10,7 @@ public class Params {
 
     String mode, coin;
     Integer grids;
-    Float budget, lower, stop_loss;
+    Double budget, lower, stop_loss;
 
     public Params() {
     }
